Use lightline, yajs. Remove vim-javascript and vim-less.
Use lightline for a better statusline.
Add lightline-pika for my personal settings.
Remove vim-javascript because there are some strange indent behavior.
Use yajs as a new syntax highlight for javascript.
Remove vim-less because I never use it...
There are other minor tweaks in this change.

Bump the version to 3.0.0 since this is a major change.